✨ useStateManagementを作成
