+++ pages/login.jsx
@@ -4,7 +4,7 @@ import Head from 'next/head'
 import { At, Book2, Key } from 'tabler-icons-react'
 import { z } from 'zod'
 import Link from 'next/link'
-import { useReducer } from 'react'
+import { useStateManagement } from 'hooks/useStateManagement'
 
 const schema = z.object({
   email: z.string().email({ message: '正しく入力されていません。' }),
@@ -13,38 +13,9 @@ const schema = z.object({
   }),
 })
 
-const initialState = {
-  loading: false,
-  data: '',
-  error: null,
-}
-
-const reducer = (state, action) => {
-  switch (action.type) {
-    case 'start':
-      return {
-        ...state,
-        loading: true,
-      }
-    case 'end':
-      return {
-        ...state,
-        loading: false,
-        data: action.data,
-      }
-    case 'error':
-      return {
-        ...state,
-        loading: false,
-        error: action.error,
-      }
-    default: {
-      throw new Error('no such action type')
-    }
-  }
-}
-
 export default function Login() {
+  const { state, dispatch } = useStateManagement()
+
   const form = useForm({
     schema: zodResolver(schema),
     initialValues: {
@@ -52,8 +23,6 @@ export default function Login() {
       password: '',
     },
   })
-
-  const [state, dispatch] = useReducer(reducer, initialState)
 
   const handleSubmit = async (values) => {
     dispatch({ type: 'start' })
